website-gen-poc/requirements.txt
langgraph>=0.0.32
langchain>=0.1.0
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0 